+++ jdbc-h2/build.gradle
@@ -13,6 +13,7 @@ dependencies {
 
     testImplementation project(':core').sourceSets.test.output
     testImplementation project(':jdbc').sourceSets.test.output
+    testImplementation project(':executor').sourceSets.test.output
     testImplementation project(':storage-local')
     testImplementation project(':tests')
 }
