+++ jdbc-mysql/build.gradle
@@ -14,6 +14,7 @@ dependencies {
 
     testImplementation project(':core').sourceSets.test.output
     testImplementation project(':jdbc').sourceSets.test.output
+    testImplementation project(':executor').sourceSets.test.output
     testImplementation project(':scheduler').sourceSets.test.output
     testImplementation project(':storage-local')
     testImplementation project(':tests')
