+++ jdbc/src/main/java/io/kestra/jdbc/JdbcWorkerJobQueueService.java
@@ -3,6 +3,7 @@ package io.kestra.jdbc;
 import io.kestra.core.exceptions.DeserializationException;
 import io.kestra.core.runners.*;
 import io.kestra.core.utils.Either;
+import io.kestra.executor.WorkerJobRunningStateStore;
 import io.kestra.jdbc.runner.JdbcQueue;
 import io.kestra.jdbc.runner.JdbcTransactionContext;
 import io.micronaut.context.ApplicationContext;
